+++ app/components/pool/details/pool-configuration.html
@@ -21,6 +21,7 @@
             <bl-text-property label="Resize timeout" [value]="decorator.resizeTimeout"></bl-text-property>
             <bl-text-property label="Task scheduling policy" [value]="decorator.taskSchedulingPolicy.nodeFillType"></bl-text-property>
             <bl-text-property label="Url" [value]="decorator.url"></bl-text-property>
+            <bl-text-property label="User accounts" [value]="decorator.userAccounts"></bl-text-property>
 		</bl-property-group>
 
 		<bl-property-group label="Cloud service configuration" class="wide" [collapsed]="true" *ngIf="pool.cloudServiceConfiguration">
