+++ app/components/pool/user-account/index.ts
@@ -1,2 +1,2 @@
-export * from "./user-accounts-picker.component";
+export * from "./user-account-picker.component";
 export * from "./user-account.module";
