+++ app/models/index.ts
@@ -54,6 +54,7 @@ export * from "./task-dependency";
 export * from "./task-execution-information";
 export * from "./task-exit-conditions";
 export * from "./task-scheduling-policy";
+export * from "./user-account";
 export * from "./virtual-machine-configuration";
 export * from "./vm-size";
 export * from "./windows-configuration";
